pyproject.toml
[tool.poetry]
name = "tinygen"
version = "0.1.0"
description = ""
authors = ["Emeka Igwegbu <emeka.igwegbu@gmail.com>"]
readme = "README.md"
packages = [{ include = "app", from = "src" }]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
GitPython = "^3.1.43"
ipykernel = "^6.29.5"
openai = "^1.51.0"
python-dotenv = "^1.0.1"
supabase = "^2.8.1"
uvicorn = "^0.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
